# src/openai/swagger.yaml
openapi: 3.0.0
info:
  title: API Documentation
  version: 1.0.0
  description: API documentation for the application
paths:
...
  /api/prompt/{id}:
    get:
      summary: Get prompt by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The ID of the prompt
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Prompt'
  /api/prompt:
    post:
      summary: Create a new prompt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Prompt'
      responses:
        '201':
          description: Prompt created successfully
...
  /api/systemprompt/{id}:
    get:
      summary: Get system prompt by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The ID of the system prompt
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemPrompt'
  /api/systemprompt:
    post:
      summary: Create a new system prompt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SystemPrompt'
      responses:
        '201':
          description: System prompt created successfully
...
  /prompts/save:
    post:
      summary: Save a prompt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Prompt'
      responses:
        '201':
          description: Prompt saved successfully
  /prompts/save/{id}:
    post:
      summary: Update a prompt by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The ID of the prompt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Prompt'
      responses:
        '201':
          description: Prompt updated successfully
...
  /prompts/edit/{id}:
    get:
      summary: Edit prompt by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The ID of the prompt
      responses:
        '200':
          description: Successful response
  /prompts/delete/{id}:
    get:
      summary: Delete prompt by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The ID of the prompt
      responses:
        '204':
          description: Prompt deleted successfully
...
  /systemprompts/save:
    post:
      summary: Save a system prompt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SystemPrompt'
      responses:
        '201':
          description: System prompt saved successfully
...
  /systemprompts/{id}:
    get:
      summary: Get system prompt by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The ID of the system prompt
      responses:
        '200':
          description: Successful response
  /systemprompts/delete/{id}:
    get:
      summary: Delete system prompt by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The ID of the system prompt
      responses:
        '204':
          description: System prompt deleted successfully
...
components:
  schemas:
    Prompt:
      type: object
      properties:
        id:
          type: string
          format: uuid
        content:
          type: string
        created_at:
          type: string
          format: date-time
...
    SystemPrompt:
      type: object
      properties:
        id:
          type: string
          format: uuid
        content:
          type: string
        created_at:
          type: string
          format: date-time
...